Update main.c
Wave interpolation is now optimised to work with fixed point and no float are used. (multiple waves are now possible)

There is now a distiction between coninuous and discontinuous interpolation (interpolate between last and first sample or not. handy for sawtooth wave).

wave structures are optimised.

(V6 from my PC. This is the last version as of right now)